Оптимизация

+++ SellerBox/ViewModels/Statistics/GroupActionsPerDayViewModel.cs
@@ -6,12 +6,12 @@ namespace SellerBox.ViewModels.Statistics
     {
         public DateTime Date { get; set; }
 
-        public long JoinGroupCount { get; set; }
-        public long LeaveGroupCount { get; set; }
-        public long BlockMessagingCount { get; set; }
-        public long AcceptMessagingCount { get; set; }
-        public long CancelMessagingCount { get; set; }
-        public long BlockedCount { get; set; }
-        public long UnblockedCount { get; set; }
+        public int JoinGroupCount { get; set; }
+        public int LeaveGroupCount { get; set; }
+        public int BlockMessagingCount { get; set; }
+        public int AcceptMessagingCount { get; set; }
+        public int CancelMessagingCount { get; set; }
+        public int BlockedCount { get; set; }
+        public int UnblockedCount { get; set; }
     }
 }
