+++ SellerBox/ViewModels/Statistics/MessagesViewModel.cs
@@ -5,7 +5,7 @@ namespace SellerBox.ViewModels.Statistics
     public class MessagesPerDayViewModel
     {
         public DateTime Date { get; set; }
-        public long Sended { get; set; }
-        public long Received { get; set; }
+        public int Sended { get; set; }
+        public int Received { get; set; }
     }
 }
