+++ SellerBox/ViewModels/Statistics/ScenariosViewModel.cs
@@ -5,6 +5,6 @@ namespace SellerBox.ViewModels.Statistics
     public class ScenariosMessagesPerDayViewModel
     {
         public DateTime Date { get; set; }
-        public long Count { get; set; }
+        public int Count { get; set; }
     }
 }
